requirements.txt
aiohttp
discord.py
Flask
python-dotenv
